requirements.txt
Flask==2.3.3
feedparser==6.0.11
flask-admin==1.6.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Limiter==3.12
Werkzeug==3.1.3
Flask-WTF==1.2.2
Flask-Migrate-4.1.0
Flask-Babel-4.0.0